+++ react/components/FilterOptionTemplate.js
@@ -51,6 +51,7 @@ const CSS_HANDLES = [
   'filter',
   'filterSelected',
   'filterAvailable',
+  'filterIsOpen',
   'filterTitle',
   'filterIcon',
   'filterContent',
@@ -194,6 +195,7 @@ const FilterOptionTemplate = ({
   const titleContainerClassName = classNames(handles.filter, 'pv5', {
     [handles.filterSelected]: selected,
     [handles.filterAvailable]: !selected,
+    [handles.filterIsOpen]: isOpen,
   })
 
   const titleClassName = classNames(