+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `"showOrderTitle"` prop to OrderBy
+- `"filterIsOpen"` CSS handle to Filter Navigator
 
 ## [3.78.1] - 2020-10-14
 ### Fixed